MapWidget: refactor, implement TRIM/MARKER/MARKER_AB/MARKER_GATE_SETS
Use common code from MapWidget.
