+++ ui/gps_gate_result.py
@@ -11,7 +11,6 @@ import pandas as pd
 from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg as FigureCanvas
 from matplotlib.figure import Figure
 from PySide6.QtCore import Qt
-from PySide6.QtWebEngineWidgets import QWebEngineView
 from PySide6.QtWidgets import (
     QCheckBox,
     QFormLayout,
@@ -35,6 +34,7 @@ from PySide6.QtWidgets import (
 )
 
 from models.virtual_elevation import VirtualElevation
+from ui.map_widget import (MapWidget, MapMode)
 
 
 class MplCanvas(FigureCanvas):
@@ -256,9 +256,12 @@ class GPSGateResult(QMainWindow):
         left_layout = QVBoxLayout(left_widget)
 
         # Map
-        if self.has_gps:
-            self.map_widget = QWebEngineView()
-            self.create_map()
+        self.map_widget = MapWidget(MapMode.MARKER_GATE_SETS, self.merged_data, self.params)
+        if self.map_widget.has_gps:
+            self.map_widget.set_trim_start(self.trim_start)
+            self.map_widget.set_trim_end(self.trim_end)
+            self.map_widget.set_gate_sets(self.gate_sets, self.detected_sections)
+            self.map_widget.update()
             left_layout.addWidget(self.map_widget, 2)
         else:
             no_gps_label = QLabel("No GPS data available")
@@ -562,7 +565,9 @@ class GPSGateResult(QMainWindow):
             self.detect_sections()
             self.calculate_ve()
             self.update_plots()
-            self.create_map()
+
+            self.map_widget.set_gate_sets(self.gate_sets, self.detected_sections)
+            self.map_widget.update()
 
             return True
         return False
@@ -595,7 +600,8 @@ class GPSGateResult(QMainWindow):
             self.detect_sections()
             self.calculate_ve()
             self.update_plots()
-            self.create_map()
+            self.map_widget.set_gate_sets(self.gate_sets, self.detected_sections)
+            self.map_widget.update()
 
     def update_calibration_end(self):
         """Update the calibration end point based on gate A's first passing"""
@@ -626,315 +632,6 @@ class GPSGateResult(QMainWindow):
                 if dir_diff < 30:  # Similar direction
                     self.calibration_end = passing["index"]
                     break
-
-    def create_map(self):
-        """Create the map showing the route with gates marked"""
-        if not self.has_gps or not self.route_points:
-            return
-
-        # Calculate center point
-        center_lat = (self.start_lat + self.end_lat) / 2
-        center_lon = (self.start_lon + self.end_lon) / 2
-
-        # Create map
-        m = folium.Map(location=[center_lat, center_lon], zoom_start=12)
-
-        try:
-            # Map time indices to route indices
-            if not hasattr(self, "route_timestamps") or not self.route_timestamps:
-                # Fall back to simple index mapping if no timestamps available
-                total_points = len(self.route_points)
-                total_records = len(self.merged_data)
-
-                if total_points > 0 and total_records > 0:
-                    route_trim_start = min(
-                        int(self.trim_start * total_points / total_records),
-                        total_points - 1,
-                    )
-                    route_trim_end = min(
-                        int(self.trim_end * total_points / total_records),
-                        total_points - 1,
-                    )
-
-                    # Map gate positions
-                    route_gate_positions = []
-                    for gate_set in self.gate_sets:
-                        gate_a = min(
-                            int(gate_set["gate_a_pos"] * total_points / total_records),
-                            total_points - 1,
-                        )
-                        gate_b = min(
-                            int(gate_set["gate_b_pos"] * total_points / total_records),
-                            total_points - 1,
-                        )
-                        route_gate_positions.append((gate_a, gate_b))
-                else:
-                    route_trim_start = 0
-                    route_trim_end = len(self.route_points) - 1
-                    route_gate_positions = []
-            else:
-                # Map using timestamps
-                route_trim_start = self.map_time_to_route_index(self.trim_start)
-                route_trim_end = self.map_time_to_route_index(self.trim_end)
-
-                # Map gate positions
-                route_gate_positions = []
-                for gate_set in self.gate_sets:
-                    gate_a = self.map_time_to_route_index(gate_set["gate_a_pos"])
-                    gate_b = self.map_time_to_route_index(gate_set["gate_b_pos"])
-                    route_gate_positions.append((gate_a, gate_b))
-
-            # Make sure indices are valid - handle None values
-            if route_trim_start is None:
-                route_trim_start = 0
-            else:
-                route_trim_start = max(0, min(route_trim_start, len(self.route_points) - 1))
-            
-            if route_trim_end is None:
-                route_trim_end = len(self.route_points) - 1
-            else:
-                route_trim_end = max(
-                    route_trim_start, min(route_trim_end, len(self.route_points) - 1)
-                )
-
-            # Clean up gate positions
-            valid_gate_positions = []
-            for gate_a, gate_b in route_gate_positions:
-                if gate_a is not None and gate_b is not None:
-                    gate_a = max(0, min(gate_a, len(self.route_points) - 1))
-                    gate_b = max(gate_a, min(gate_b, len(self.route_points) - 1))
-                    valid_gate_positions.append((gate_a, gate_b))
-
-            # 1. Draw the parts before trim_start with dashed blue line (if exists)
-            if route_trim_start > 0:
-                pre_trim_route = self.route_points[: route_trim_start + 1]
-                folium.PolyLine(
-                    pre_trim_route,
-                    color="#4363d8",  # Blue color
-                    weight=3,
-                    opacity=0.5,
-                    dash_array="5,10",  # Dashed line
-                    popup="Pre-selected portion",
-                ).add_to(m)
-
-            # 2. Draw the parts after trim_end with dashed blue line (if exists)
-            if route_trim_end < len(self.route_points) - 1:
-                post_trim_route = self.route_points[route_trim_end:]
-                folium.PolyLine(
-                    post_trim_route,
-                    color="#4363d8",  # Blue color
-                    weight=3,
-                    opacity=0.5,
-                    dash_array="5,10",  # Dashed line
-                    popup="Post-selected portion",
-                ).add_to(m)
-
-            # 3. Draw the selected portion with solid blue line
-            trimmed_route = self.route_points[route_trim_start : route_trim_end + 1]
-            if len(trimmed_route) > 1:
-                folium.PolyLine(
-                    trimmed_route,
-                    color="#4363d8",  # Blue color
-                    weight=5,  # Slightly thicker
-                    opacity=1.0,  # Full opacity
-                    popup="Selected portion",
-                ).add_to(m)
-
-                # Add trim markers
-                folium.Marker(
-                    location=trimmed_route[0],
-                    popup="Trim Start",
-                    icon=folium.Icon(color="green", icon="play", prefix="fa"),
-                ).add_to(m)
-
-                folium.Marker(
-                    location=trimmed_route[-1],
-                    popup="Trim End",
-                    icon=folium.Icon(color="red", icon="stop", prefix="fa"),
-                ).add_to(m)
-
-                # Add gate markers
-                gate_colors = [
-                    "blue",
-                    "purple",
-                    "orange",
-                    "cadetblue",
-                    "darkred",
-                    "darkgreen",
-                ]
-
-                for i, (gate_a_pos, gate_b_pos) in enumerate(valid_gate_positions):
-                    color = gate_colors[i % len(gate_colors)]
-                    gate_set_idx = i
-
-                    # Gate A marker
-                    if 0 <= gate_a_pos - route_trim_start < len(trimmed_route):
-                        gate_a_loc = trimmed_route[gate_a_pos - route_trim_start]
-
-                        # Use CircleMarker for Gate A
-                        folium.CircleMarker(
-                            location=gate_a_loc,
-                            radius=8,
-                            color=color,
-                            fill=True,
-                            fill_color=color,
-                            fill_opacity=0.8,
-                            popup=f"Gate {gate_set_idx+1}A",
-                            weight=2,
-                        ).add_to(m)
-
-                        # Add label
-                        folium.map.Marker(
-                            gate_a_loc,
-                            icon=folium.DivIcon(
-                                icon_size=(20, 20),
-                                icon_anchor=(10, 10),
-                                html=f'<div style="font-size: 12pt; font-weight: bold; color: white; background-color: {color}; border-radius: 50%; width: 20px; height: 20px; line-height: 20px; text-align: center;">{gate_set_idx+1}A</div>',
-                            ),
-                        ).add_to(m)
-
-                        # Add detection radius
-                        folium.Circle(
-                            location=gate_a_loc,
-                            radius=20,  # 20 meters
-                            color=color,
-                            fill=True,
-                            fill_color=color,
-                            fill_opacity=0.2,
-                            popup=f"Gate {gate_set_idx+1}A Detection Zone",
-                        ).add_to(m)
-
-                    # Gate B marker
-                    if 0 <= gate_b_pos - route_trim_start < len(trimmed_route):
-                        gate_b_loc = trimmed_route[gate_b_pos - route_trim_start]
-
-                        # Use CircleMarker for Gate B
-                        folium.CircleMarker(
-                            location=gate_b_loc,
-                            radius=8,
-                            color=color,
-                            fill=True,
-                            fill_color=color,
-                            fill_opacity=0.8,
-                            popup=f"Gate {gate_set_idx+1}B",
-                            weight=2,
-                        ).add_to(m)
-
-                        # Add label
-                        folium.map.Marker(
-                            gate_b_loc,
-                            icon=folium.DivIcon(
-                                icon_size=(20, 20),
-                                icon_anchor=(10, 10),
-                                html=f'<div style="font-size: 12pt; font-weight: bold; color: white; background-color: {color}; border-radius: 50%; width: 20px; height: 20px; line-height: 20px; text-align: center;">{gate_set_idx+1}B</div>',
-                            ),
-                        ).add_to(m)
-
-                        # Add detection radius
-                        folium.Circle(
-                            location=gate_b_loc,
-                            radius=20,  # 20 meters
-                            color=color,
-                            fill=True,
-                            fill_color=color,
-                            fill_opacity=0.2,
-                            popup=f"Gate {gate_set_idx+1}B Detection Zone",
-                        ).add_to(m)
-
-                # Highlight detected sections if available
-                section_colors = {
-                    0: "#ff7f0e",  # Orange
-                    1: "#2ca02c",  # Green
-                    2: "#d62728",  # Red
-                    3: "#9467bd",  # Purple
-                    4: "#8c564b",  # Brown
-                    5: "#e377c2",  # Pink
-                }
-
-                for i, section in enumerate(self.detected_sections):
-                    gate_idx = section.get("gate_set", 0)
-                    section_color = section_colors.get(
-                        gate_idx, "#1f77b4"
-                    )  # Default blue
-
-                    # Section: A to B
-                    if "start_idx" in section and "end_idx" in section:
-                        section_start = self.map_time_to_route_index(
-                            section["start_idx"]
-                        )
-                        section_end = self.map_time_to_route_index(section["end_idx"])
-
-                        if section_start is not None and section_end is not None:
-                            section_start = max(
-                                0, min(section_start, len(self.route_points) - 1)
-                            )
-                            section_end = max(
-                                section_start,
-                                min(section_end, len(self.route_points) - 1),
-                            )
-
-                            section_route = self.route_points[
-                                section_start : section_end + 1
-                            ]
-                            if len(section_route) > 1:
-                                folium.PolyLine(
-                                    section_route,
-                                    color=section_color,
-                                    weight=5,
-                                    opacity=0.7,
-                                    popup=f"Section {section.get('section_id', i+1)}: Gate {gate_idx+1}A → {gate_idx+1}B",
-                                ).add_to(m)
-
-        except Exception as e:
-            print(f"Error highlighting route: {e}")
-
-        # Calculate bounds for automatic zoom
-        try:
-            if self.route_points:
-                lats = [p[0] for p in self.route_points]
-                lons = [p[1] for p in self.route_points]
-                min_lat, max_lat = min(lats), max(lats)
-                min_lon, max_lon = min(lons), max(lons)
-
-                # Add some padding (5%)
-                lat_padding = (max_lat - min_lat) * 0.05
-                lon_padding = (max_lon - min_lon) * 0.05
-                bounds = [
-                    [min_lat - lat_padding, min_lon - lon_padding],
-                    [max_lat + lat_padding, max_lon + lon_padding],
-                ]
-                m.fit_bounds(bounds)
-        except Exception as e:
-            print(f"Error fitting map bounds: {e}")
-
-        # Add wind arrow AFTER map bounds have been set
-        wind_speed = self.params.get("wind_speed")
-        wind_dir = self.params.get("wind_direction")
-
-        if wind_speed not in [None, 0] and wind_dir is not None:
-            # Create a custom HTML element for the wind arrow
-            wind_html = f"""
-            <div id="wind-arrow" style="position: absolute; top: 10px; right: 10px; 
-                    background-color: rgba(255, 255, 255, 0.9); padding: 10px; 
-                    border-radius: 5px; border: 1px solid #4363d8; z-index: 1000;
-                    box-shadow: 0 2px 5px rgba(0,0,0,0.2);">
-                <div style="font-weight: bold; text-align: center; margin-bottom: 5px; color: #4363d8;">Wind</div>
-                <div style="text-align: center;">
-                    <div style="font-size: 28px; transform: rotate({wind_dir + 180}deg); color: #4363d8;">↑</div>
-                    <div style="font-size: 12px; margin-top: 5px;">{wind_speed} m/s</div>
-                    <div style="font-size: 12px;">{wind_dir}°</div>
-                </div>
-            </div>
-            """
-
-            # Add the HTML element to the map
-            m.get_root().html.add_child(folium.Element(wind_html))
-
-        # Save map to HTML and load into QWebEngineView
-        data = io.BytesIO()
-        m.save(data, close_file=False)
-        html_content = data.getvalue().decode()
-        self.map_widget.setHtml(html_content)
 
     def update_config_text(self):
         """Update the configuration text display"""
@@ -1730,7 +1427,8 @@ class GPSGateResult(QMainWindow):
         self.update_plots()
 
         # Update map
-        self.create_map()
+        self.map_widget.set_gate_sets(self.gate_sets, self.detected_sections)
+        self.map_widget.update()
 
         # Update control min/max values
         self.update_gate_controls()
@@ -1770,7 +1468,8 @@ class GPSGateResult(QMainWindow):
         self.update_plots()
 
         # Update map
-        self.create_map()
+        self.map_widget.set_gate_sets(self.gate_sets, self.detected_sections)
+        self.map_widget.update()
 
         # Update control min/max values
         self.update_gate_controls()
@@ -1810,7 +1509,9 @@ class GPSGateResult(QMainWindow):
         self.update_plots()
 
         # Update map to show trim points
-        self.create_map()
+        self.map_widget.set_trim_start(self.trim_start)
+        self.map_widget.set_gate_sets(self.gate_sets, self.detected_sections)
+        self.map_widget.update()
 
         # Update all gate controls
         self.update_gate_controls()
@@ -1850,7 +1551,9 @@ class GPSGateResult(QMainWindow):
         self.update_plots()
 
         # Update map to show trim points
-        self.create_map()
+        self.map_widget.set_trim_end(self.trim_end)
+        self.map_widget.set_gate_sets(self.gate_sets, self.detected_sections)
+        self.map_widget.update()
 
         # Update all gate controls
         self.update_gate_controls()
@@ -2073,38 +1776,3 @@ class GPSGateResult(QMainWindow):
         self.analysis_window = AnalysisWindow(self.fit_file, self.settings)
         self.analysis_window.show()
         self.close()
-
-    def map_time_to_route_index(self, time_index):
-        """
-        Map a time index from the full dataset to the corresponding index in route_points
-
-        Parameters:
-        -----------
-        time_index : int
-            Index in the merged_data dataframe
-
-        Returns:
-        --------
-        int
-            Corresponding index in the route_points list, or None if not mappable
-        """
-        if not hasattr(self, "route_timestamps") or not self.route_timestamps:
-            return None
-
-        if time_index < 0 or time_index >= len(self.merged_data):
-            return None
-
-        # Get the timestamp at this index
-        target_timestamp = self.merged_data["timestamp"].iloc[time_index]
-
-        # Find the closest timestamp in route_timestamps
-        if target_timestamp in self.route_timestamps:
-            return self.route_timestamps.index(target_timestamp)
-
-        # If not found directly, find the closest one
-        for i, ts in enumerate(self.route_timestamps):
-            if ts >= target_timestamp:
-                return i
-
-        # If we get here, target_timestamp is after all route_timestamps
-        return len(self.route_timestamps) - 1
