+++ ui/out_and_back_result.py
@@ -9,7 +9,6 @@ import numpy as np
 from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg as FigureCanvas
 from matplotlib.figure import Figure
 from PySide6.QtCore import Qt
-from PySide6.QtWebEngineWidgets import QWebEngineView
 from PySide6.QtWidgets import (
     QCheckBox,
     QFormLayout,
@@ -32,6 +31,7 @@ from PySide6.QtWidgets import (
 )
 
 from models.virtual_elevation import VirtualElevation
+from ui.map_widget import (MapWidget, MapMode)
 
 
 class MplCanvas(FigureCanvas):
@@ -216,9 +216,13 @@ class OutAndBackResult(QMainWindow):
         left_layout = QVBoxLayout(left_widget)
 
         # Map
-        if self.has_gps:
-            self.map_widget = QWebEngineView()
-            self.create_map()
+        self.map_widget = MapWidget(MapMode.MARKER_AB, self.merged_data, self.params)
+        if self.map_widget:
+            self.map_widget.set_marker_a_pos(self.gps_marker_a_pos)
+            self.map_widget.set_marker_b_pos(self.gps_marker_b_pos)
+            self.map_widget.set_trim_start(self.trim_start)
+            self.map_widget.set_trim_end(self.trim_end)
+            self.map_widget.update()
             left_layout.addWidget(self.map_widget, 2)
         else:
             no_gps_label = QLabel("No GPS data available")
@@ -391,348 +395,6 @@ class OutAndBackResult(QMainWindow):
 
         main_widget.setLayout(main_layout)
         self.setCentralWidget(main_widget)
-
-    def create_map(self):
-        """Create the map showing the merged track with start/end markers, GPS markers, and trimmed portion"""
-        if not self.has_gps or not self.route_points:
-            return
-
-        # Calculate center point
-        center_lat = (self.start_lat + self.end_lat) / 2
-        center_lon = (self.start_lon + self.end_lon) / 2
-
-        # Create map
-        m = folium.Map(location=[center_lat, center_lon], zoom_start=12)
-
-        try:
-            # Map time indices to route indices
-            if not hasattr(self, "route_timestamps") or not self.route_timestamps:
-                # Fall back to simple index mapping if no timestamps available
-                total_points = len(self.route_points)
-                total_records = len(self.merged_data)
-
-                if total_points > 0 and total_records > 0:
-                    route_trim_start = min(
-                        int(self.trim_start * total_points / total_records),
-                        total_points - 1,
-                    )
-                    route_trim_end = min(
-                        int(self.trim_end * total_points / total_records),
-                        total_points - 1,
-                    )
-                    route_marker_a_pos = min(
-                        int(self.gps_marker_a_pos * total_points / total_records),
-                        total_points - 1,
-                    )
-                    route_marker_b_pos = min(
-                        int(self.gps_marker_b_pos * total_points / total_records),
-                        total_points - 1,
-                    )
-                else:
-                    route_trim_start = 0
-                    route_trim_end = len(self.route_points) - 1
-                    route_marker_a_pos = int((route_trim_start + route_trim_end) * 0.25)
-                    route_marker_b_pos = int((route_trim_start + route_trim_end) * 0.75)
-            else:
-                # Map using timestamps
-                route_trim_start = self.map_time_to_route_index(self.trim_start)
-                route_trim_end = self.map_time_to_route_index(self.trim_end)
-                route_marker_a_pos = self.map_time_to_route_index(self.gps_marker_a_pos)
-                route_marker_b_pos = self.map_time_to_route_index(self.gps_marker_b_pos)
-
-                # Fall back if mapping fails
-                if (
-                    route_trim_start is None
-                    or route_trim_end is None
-                    or route_marker_a_pos is None
-                    or route_marker_b_pos is None
-                ):
-                    total_points = len(self.route_points)
-                    total_records = len(self.merged_data)
-
-                    if total_points > 0 and total_records > 0:
-                        route_trim_start = min(
-                            int(self.trim_start * total_points / total_records),
-                            total_points - 1,
-                        )
-                        route_trim_end = min(
-                            int(self.trim_end * total_points / total_records),
-                            total_points - 1,
-                        )
-                        route_marker_a_pos = min(
-                            int(self.gps_marker_a_pos * total_points / total_records),
-                            total_points - 1,
-                        )
-                        route_marker_b_pos = min(
-                            int(self.gps_marker_b_pos * total_points / total_records),
-                            total_points - 1,
-                        )
-                    else:
-                        route_trim_start = 0
-                        route_trim_end = len(self.route_points) - 1
-                        route_marker_a_pos = int(
-                            (route_trim_start + route_trim_end) * 0.25
-                        )
-                        route_marker_b_pos = int(
-                            (route_trim_start + route_trim_end) * 0.75
-                        )
-
-            # Make sure indices are valid
-            route_trim_start = max(0, min(route_trim_start, len(self.route_points) - 1))
-            route_trim_end = max(
-                route_trim_start, min(route_trim_end, len(self.route_points) - 1)
-            )
-            route_marker_a_pos = max(
-                route_trim_start, min(route_marker_a_pos, route_trim_end)
-            )
-            route_marker_b_pos = max(
-                route_trim_start, min(route_marker_b_pos, route_trim_end)
-            )
-
-            # 1. Draw the parts before trim_start with dashed blue line (if exists)
-            if route_trim_start > 0:
-                pre_trim_route = self.route_points[: route_trim_start + 1]
-                folium.PolyLine(
-                    pre_trim_route,
-                    color="#4363d8",  # Blue color
-                    weight=3,
-                    opacity=0.5,
-                    dash_array="5,10",  # Dashed line
-                    popup="Pre-selected portion",
-                ).add_to(m)
-
-            # 2. Draw the parts after trim_end with dashed blue line (if exists)
-            if route_trim_end < len(self.route_points) - 1:
-                post_trim_route = self.route_points[route_trim_end:]
-                folium.PolyLine(
-                    post_trim_route,
-                    color="#4363d8",  # Blue color
-                    weight=3,
-                    opacity=0.5,
-                    dash_array="5,10",  # Dashed line
-                    popup="Post-selected portion",
-                ).add_to(m)
-
-            # 3. Draw the selected portion with solid blue line
-            trimmed_route = self.route_points[route_trim_start : route_trim_end + 1]
-            if len(trimmed_route) > 1:
-                folium.PolyLine(
-                    trimmed_route,
-                    color="#4363d8",  # Blue color
-                    weight=5,  # Slightly thicker
-                    opacity=1.0,  # Full opacity
-                    popup="Selected portion",
-                ).add_to(m)
-
-                # Add trim markers
-                folium.Marker(
-                    location=trimmed_route[0],
-                    popup="Trim Start",
-                    icon=folium.Icon(color="green", icon="play", prefix="fa"),
-                ).add_to(m)
-
-                folium.Marker(
-                    location=trimmed_route[-1],
-                    popup="Trim End",
-                    icon=folium.Icon(color="red", icon="stop", prefix="fa"),
-                ).add_to(m)
-
-            if 0 <= route_marker_a_pos - route_trim_start < len(trimmed_route):
-                marker_a_location = trimmed_route[route_marker_a_pos - route_trim_start]
-
-                # Use a CircleMarker instead of standard marker
-                folium.CircleMarker(
-                    location=marker_a_location,
-                    radius=8,
-                    color="#3186cc",
-                    fill=True,
-                    fill_color="#3186cc",
-                    fill_opacity=0.8,
-                    popup="Point A",
-                    weight=2,
-                ).add_to(m)
-
-                # Add a letter label using DivIcon
-                folium.map.Marker(
-                    marker_a_location,
-                    icon=folium.DivIcon(
-                        icon_size=(20, 20),
-                        icon_anchor=(10, 10),
-                        html='<div style="font-size: 12pt; font-weight: bold; color: white; background-color: #3186cc; border-radius: 50%; width: 20px; height: 20px; line-height: 20px; text-align: center;">A</div>',
-                    ),
-                ).add_to(m)
-
-                # Add detection radius
-                folium.Circle(
-                    location=marker_a_location,
-                    radius=20,  # 20 meters
-                    color="#3186cc",
-                    fill=True,
-                    fill_color="#3186cc",
-                    fill_opacity=0.2,
-                    popup="Point A Detection Zone",
-                ).add_to(m)
-
-            # Replace marker B implementation
-            if 0 <= route_marker_b_pos - route_trim_start < len(trimmed_route):
-                marker_b_location = trimmed_route[route_marker_b_pos - route_trim_start]
-
-                # Use a CircleMarker instead of standard marker
-                folium.CircleMarker(
-                    location=marker_b_location,
-                    radius=8,
-                    color="#9c27b0",
-                    fill=True,
-                    fill_color="#9c27b0",
-                    fill_opacity=0.8,
-                    popup="Point B",
-                    weight=2,
-                ).add_to(m)
-
-                # Add a letter label using DivIcon
-                folium.map.Marker(
-                    marker_b_location,
-                    icon=folium.DivIcon(
-                        icon_size=(20, 20),
-                        icon_anchor=(10, 10),
-                        html='<div style="font-size: 12pt; font-weight: bold; color: white; background-color: #9c27b0; border-radius: 50%; width: 20px; height: 20px; line-height: 20px; text-align: center;">B</div>',
-                    ),
-                ).add_to(m)
-
-                # Add detection radius
-                folium.Circle(
-                    location=marker_b_location,
-                    radius=20,  # 20 meters
-                    color="#9c27b0",
-                    fill=True,
-                    fill_color="#9c27b0",
-                    fill_opacity=0.2,
-                    popup="Point B Detection Zone",
-                ).add_to(m)
-
-                # 4. Highlight detected sections if available
-                if hasattr(self, "detected_sections") and self.detected_sections:
-                    for i, section in enumerate(self.detected_sections):
-                        # Use different colors for outbound and inbound parts
-                        outbound_color = "#ff7f0e"  # Orange
-                        inbound_color = "#2ca02c"  # Green
-
-                        # Outbound: A to B
-                        if (
-                            "outbound_start_idx" in section
-                            and "outbound_end_idx" in section
-                        ):
-                            outbound_start = self.map_time_to_route_index(
-                                section["outbound_start_idx"]
-                            )
-                            outbound_end = self.map_time_to_route_index(
-                                section["outbound_end_idx"]
-                            )
-
-                            if outbound_start is not None and outbound_end is not None:
-                                outbound_start = max(
-                                    0, min(outbound_start, len(self.route_points) - 1)
-                                )
-                                outbound_end = max(
-                                    outbound_start,
-                                    min(outbound_end, len(self.route_points) - 1),
-                                )
-
-                                outbound_route = self.route_points[
-                                    outbound_start : outbound_end + 1
-                                ]
-                                if len(outbound_route) > 1:
-                                    folium.PolyLine(
-                                        outbound_route,
-                                        color=outbound_color,
-                                        weight=5,
-                                        opacity=0.7,
-                                        popup=f"Section {i+1} Outbound (A→B)",
-                                    ).add_to(m)
-
-                        # Inbound: B to A
-                        if (
-                            "inbound_start_idx" in section
-                            and "inbound_end_idx" in section
-                        ):
-                            inbound_start = self.map_time_to_route_index(
-                                section["inbound_start_idx"]
-                            )
-                            inbound_end = self.map_time_to_route_index(
-                                section["inbound_end_idx"]
-                            )
-
-                            if inbound_start is not None and inbound_end is not None:
-                                inbound_start = max(
-                                    0, min(inbound_start, len(self.route_points) - 1)
-                                )
-                                inbound_end = max(
-                                    inbound_start,
-                                    min(inbound_end, len(self.route_points) - 1),
-                                )
-
-                                inbound_route = self.route_points[
-                                    inbound_start : inbound_end + 1
-                                ]
-                                if len(inbound_route) > 1:
-                                    folium.PolyLine(
-                                        inbound_route,
-                                        color=inbound_color,
-                                        weight=5,
-                                        opacity=0.7,
-                                        popup=f"Section {i+1} Inbound (B→A)",
-                                    ).add_to(m)
-
-        except Exception as e:
-            print(f"Error highlighting route: {e}")
-
-        # Calculate bounds for automatic zoom
-        try:
-            if self.route_points:
-                lats = [p[0] for p in self.route_points]
-                lons = [p[1] for p in self.route_points]
-                min_lat, max_lat = min(lats), max(lats)
-                min_lon, max_lon = min(lons), max(lons)
-
-                # Add some padding (5%)
-                lat_padding = (max_lat - min_lat) * 0.05
-                lon_padding = (max_lon - min_lon) * 0.05
-                bounds = [
-                    [min_lat - lat_padding, min_lon - lon_padding],
-                    [max_lat + lat_padding, max_lon + lon_padding],
-                ]
-                m.fit_bounds(bounds)
-        except Exception as e:
-            print(f"Error fitting map bounds: {e}")
-
-        # Add wind arrow AFTER map bounds have been set
-        wind_speed = self.params.get("wind_speed")
-        wind_dir = self.params.get("wind_direction")
-
-        if wind_speed not in [None, 0] and wind_dir is not None:
-            # Create a custom HTML element for the wind arrow
-            wind_html = f"""
-            <div id="wind-arrow" style="position: absolute; top: 10px; right: 10px; 
-                    background-color: rgba(255, 255, 255, 0.9); padding: 10px; 
-                    border-radius: 5px; border: 1px solid #4363d8; z-index: 1000;
-                    box-shadow: 0 2px 5px rgba(0,0,0,0.2);">
-                <div style="font-weight: bold; text-align: center; margin-bottom: 5px; color: #4363d8;">Wind</div>
-                <div style="text-align: center;">
-                    <div style="font-size: 28px; transform: rotate({wind_dir + 180}deg); color: #4363d8;">↑</div>
-                    <div style="font-size: 12px; margin-top: 5px;">{wind_speed} m/s</div>
-                    <div style="font-size: 12px;">{wind_dir}°</div>
-                </div>
-            </div>
-            """
-
-            # Add the HTML element to the map
-            m.get_root().html.add_child(folium.Element(wind_html))
-
-        # Save map to HTML and load into QWebEngineView
-        data = io.BytesIO()
-        m.save(data, close_file=False)
-        html_content = data.getvalue().decode()
-        self.map_widget.setHtml(html_content)
 
     def update_config_text(self):
         """Update the configuration text display"""
@@ -1668,11 +1330,13 @@ class OutAndBackResult(QMainWindow):
             self.gps_marker_a_slider.setValue(value)
             self.gps_marker_a_pos = value
             self.gps_marker_a_label.setText(f"{value} s")
+            self.map_widget.set_marker_a_pos(self.gps_marker_a_pos)
 
         if self.gps_marker_b_pos < value:
             self.gps_marker_b_slider.setValue(value)
             self.gps_marker_b_pos = value
             self.gps_marker_b_label.setText(f"{value} s")
+            self.map_widget.set_marker_b_pos(self.gps_marker_b_pos)
 
         # Detect sections based on new trim values
         self.detect_sections()
@@ -1682,7 +1346,8 @@ class OutAndBackResult(QMainWindow):
         self.update_plots()
 
         # Update map to show trim points
-        self.create_map()
+        self.map_widget.set_trim_start(self.trim_start)
+        self.map_widget.update()
 
     def on_trim_end_changed(self, value):
         """Handle trim end slider value change"""
@@ -1703,11 +1368,13 @@ class OutAndBackResult(QMainWindow):
             self.gps_marker_a_slider.setValue(value)
             self.gps_marker_a_pos = value
             self.gps_marker_a_label.setText(f"{value} s")
+            self.map_widget.set_marker_a_pos(self.gps_marker_a_pos)
 
         if self.gps_marker_b_pos > value:
             self.gps_marker_b_slider.setValue(value)
             self.gps_marker_b_pos = value
             self.gps_marker_b_label.setText(f"{value} s")
+            self.map_widget.set_marker_b_pos(self.gps_marker_b_pos)
 
         # Detect sections based on new trim values
         self.detect_sections()
@@ -1717,12 +1384,14 @@ class OutAndBackResult(QMainWindow):
         self.update_plots()
 
         # Update map to show trim points
-        self.create_map()
+        self.map_widget.set_trim_end(self.trim_end)
+        self.map_widget.update()
 
     def on_gps_marker_a_changed(self, value):
         """Handle GPS marker A slider value change"""
         self.gps_marker_a_pos = value
         self.gps_marker_a_label.setText(f"{value} s")
+        self.map_widget.set_marker_a_pos(self.gps_marker_a_pos)
 
         # Detect sections based on new marker position
         self.detect_sections()
@@ -1732,12 +1401,13 @@ class OutAndBackResult(QMainWindow):
         self.update_plots()
 
         # Update map to show markers
-        self.create_map()
+        self.map_widget.update()
 
     def on_gps_marker_b_changed(self, value):
         """Handle GPS marker B slider value change"""
         self.gps_marker_b_pos = value
         self.gps_marker_b_label.setText(f"{value} s")
+        self.map_widget.set_marker_b_pos(self.gps_marker_b_pos)
 
         # Detect sections based on new marker position
         self.detect_sections()
@@ -1747,7 +1417,7 @@ class OutAndBackResult(QMainWindow):
         self.update_plots()
 
         # Update map to show markers
-        self.create_map()
+        self.map_widget.update()
 
     def on_cda_changed(self, value):
         """Handle CdA slider value change"""
@@ -1949,38 +1619,3 @@ class OutAndBackResult(QMainWindow):
         self.analysis_window = AnalysisWindow(self.fit_file, self.settings)
         self.analysis_window.show()
         self.close()
-
-    def map_time_to_route_index(self, time_index):
-        """
-        Map a time index from the full dataset to the corresponding index in route_points
-
-        Parameters:
-        -----------
-        time_index : int
-            Index in the merged_data dataframe
-
-        Returns:
-        --------
-        int
-            Corresponding index in the route_points list, or None if not mappable
-        """
-        if not hasattr(self, "route_timestamps") or not self.route_timestamps:
-            return None
-
-        if time_index < 0 or time_index >= len(self.merged_data):
-            return None
-
-        # Get the timestamp at this index
-        target_timestamp = self.merged_data["timestamp"].iloc[time_index]
-
-        # Find the closest timestamp in route_timestamps
-        if target_timestamp in self.route_timestamps:
-            return self.route_timestamps.index(target_timestamp)
-
-        # If not found directly, find the closest one
-        for i, ts in enumerate(self.route_timestamps):
-            if ts >= target_timestamp:
-                return i
-
-        # If we get here, target_timestamp is after all route_timestamps
-        return len(self.route_timestamps) - 1
